[REFACTOR] Moving common functions to utils/ dir
Common functions that were previously in app/utils.py have been moved
to app/utils/common.py.

This is to make the codebase more organized and easier to maintain.
